+++ dashboard/src/scenes/person/components/PersonTeamHistory.tsx
@@ -1,0 +1,166 @@
+import { formatDateWithFullMonth, dayjsInstance } from "../../../services/date";
+import { borderColors, teamsColors } from "../../../components/TagTeam";
+import { TeamInstance } from "../../../types/team";
+import { PersonHistoryEntry, PersonInstance } from "../../../types/person";
+
+interface TeamHistorySlice {
+  team: string;
+  startDate: string | Date;
+  endDate: string | Date;
+}
+
+function getPersonTeamHistory(changes: Array<PersonHistoryEntry>, creationDate: string | Date): Array<TeamHistorySlice> {
+  const teamHistory = {};
+  const result: Array<TeamHistorySlice> = [];
+
+  changes = changes.filter((change) => change.data && change.data.assignedTeams);
+  changes.sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime());
+
+  if (changes.length === 0) return [];
+
+  if (changes.length > 0) {
+    const initialTeams = changes[0].data.assignedTeams.oldValue;
+    initialTeams.forEach((team) => {
+      teamHistory[team] = { startDate: creationDate, endDate: null };
+    });
+  }
+
+  for (const change of changes) {
+    const date = change.date;
+    const oldTeams = change.data.assignedTeams.oldValue;
+    const newTeams = change.data.assignedTeams.newValue;
+
+    // Équipes quittées
+    for (const team of oldTeams) {
+      if (!newTeams.includes(team)) {
+        if (teamHistory[team] && !teamHistory[team].endDate) {
+          teamHistory[team].endDate = date;
+          result.push({ team, startDate: teamHistory[team].startDate, endDate: new Date(date).toISOString() });
+          delete teamHistory[team];
+        }
+      }
+    }
+
+    // Équipes rejointes
+    for (const team of newTeams) {
+      if (!oldTeams.includes(team)) {
+        if (!teamHistory[team]) {
+          teamHistory[team] = { startDate: date, endDate: null };
+        }
+      }
+    }
+  }
+
+  // Fermer les périodes ouvertes pour les équipes toujours actives
+  const today = new Date().toISOString();
+  for (const team in teamHistory) {
+    if (teamHistory[team].startDate && !teamHistory[team].endDate) {
+      result.push({ team, startDate: teamHistory[team].startDate, endDate: today });
+    }
+  }
+
+  return result;
+}
+
+const GanttChart = ({ data, teams }: { data: Array<TeamHistorySlice>; teams: Array<TeamInstance> }) => {
+  const globalStartDate = new Date(Math.min(...data.map((item) => new Date(item.startDate).getTime())));
+  const globalEndDate = new Date(Math.max(...data.map((item) => new Date(item.endDate).getTime())));
+  const totalDuration = globalEndDate.getTime() - globalStartDate.getTime();
+
+  const calculatePosition = (startDate: Date | string, endDate: Date | string, totalWidth: number) => {
+    const start = new Date(startDate);
+    const end = new Date(endDate);
+    const startOffset = start.getTime() - globalStartDate.getTime();
+    const duration = end.getTime() - start.getTime();
+    const width = (duration / totalDuration) * totalWidth;
+    const position = (startOffset / totalDuration) * totalWidth;
+    return { width, position };
+  };
+
+  const groupedData = data.reduce((acc, item) => {
+    if (!acc[item.team]) {
+      acc[item.team] = [];
+    }
+    acc[item.team].push(item);
+    return acc;
+  }, {});
+
+  const dataForDisplay = [];
+  let lineIndex = 0;
+  let hasSkipped = false;
+
+  for (const teamId in groupedData) {
+    let found = false;
+    for (const item of groupedData[teamId]) {
+      const teamIndex = teams?.findIndex((t) => t._id === item.team);
+      const { width, position } = calculatePosition(item.startDate, item.endDate, 800);
+      if (width < 10) {
+        hasSkipped = true;
+        continue; // Ne pas afficher les éléments trop courts (10px minimum)
+      }
+      found = true;
+      dataForDisplay.push({
+        ...item,
+        width,
+        position,
+        top: lineIndex * 38,
+        backgroundColor: teamsColors[teamIndex % teamsColors?.length],
+        borderColor: borderColors[teamIndex % borderColors?.length],
+      });
+    }
+    if (found) lineIndex++;
+  }
+
+  if (dataForDisplay.length === 0) return null;
+
+  return (
+    <div className="tw-border tw-border-gray-200 tw-rounded-lg tw-shadow tw-overflow-x-auto tw-overflow-y-hidden tw-p-2 tw-max-w-[824px] tw-mx-auto">
+      <h3 className="tw-mb-4 tw-pb-1 tw-text-lg tw-font-semibold tw-border-b tw-border-zinc-200">
+        Mouvements d'équipe du {formatDateWithFullMonth(globalStartDate)} au {formatDateWithFullMonth(globalEndDate)}
+      </h3>
+      <div className="tw-relative tw-w-[800px] tw-overflow-hidden" style={{ height: `${lineIndex * 38}px` }}>
+        {dataForDisplay.map((item, index) => {
+          const teamName = teams.find((t) => t._id === item.team)?.name || "n/c";
+          return (
+            <div
+              key={index}
+              className="tw-absolute tw-h-8 tw-flex tw-flex-col tw-rounded-sm tw-items-start"
+              style={{
+                width: `${item.width}px`,
+                left: `${item.position}px`,
+                top: `${item.top}px`,
+                backgroundColor: item.backgroundColor,
+                border: "1px solid " + item.borderColor,
+              }}
+              title={teamName}
+            >
+              <div className="text-white tw-text-xs tw-font-semibold tw-truncate tw-pl-1">{teamName}</div>
+              <div className="text-white tw-text-[10px] tw-truncate tw-pl-1 tw-leading-3">
+                {dayjsInstance(item.startDate).format("DD/MM/YY")} au {dayjsInstance(item.endDate).format("DD/MM/YY")}
+              </div>
+            </div>
+          );
+        })}
+      </div>
+      {hasSkipped ? (
+        <div className="tw-italic tw-text-xs">
+          Plusieurs mouvements d'équipes sur des temps courts sont invisibles à cette échelle, consultez l'historique pour le détail complet.
+        </div>
+      ) : null}
+    </div>
+  );
+};
+
+export default function PersonTeamHistory({
+  person,
+  history,
+  teams,
+}: {
+  person: PersonInstance;
+  history: Array<PersonHistoryEntry>;
+  teams: Array<TeamInstance>;
+}) {
+  const data = getPersonTeamHistory(history || [], person.createdAt);
+
+  return <GanttChart data={data} teams={teams} />;
+}
